Add support for SwiftLint config
Why:

* People want to be able to configure SwiftLint to their liking

This change addresses the need by:

* Passing through the user config to Swiftlint
* Creating `bin/hound-swiftlint` that we call to call out
  to the swiftlint binary on the system.
  - This allows us to run multiple shell commands without having
    to type them all out in ruby.
  - It also makes testing easier because we can make simpler assertions.
* Dependency injecting a `SystemCall` so that we can swap it out
  in test and not actually call the system

Also:

Write to disk instead of $STDIN when calling SwiftLint

$STDIN is pretty limited and we needed to be able
to give it a file name so that the `exclude` config
works properly.

Rather than implementing some sort of file naming
API that would have to get merged into SwiftLint
it was easier to write to disk and probably
more future proof because this is how people
will typically be using it.

Because reading from a path causes more clang
output I also improved the SwiftLint::Violation
class's regex to be able to determine if it is
parsable to remove those lines from linted output.